refactor: cross-shard tidy-up — dead fallback, generic frechet_variance, canonical imports
- compute_pairwise_distances: remove the dead legacy module-API (_dist)
  fallback — manifold_module is typed Manifold, whose protocol requires
  dist, and nothing in the library exposes a bare _dist module API
- frechet_variance: retype manifold: Poincare -> Manifold; the gyro
  normalization layers already call it with Hyperboloid and
  ProperVelocity instances
- hyperboloid_linear/busemann_core: import MIN_NORM from its canonical
  home hyperbolix.utils.math_utils instead of the manifolds.hyperboloid
  re-export
- euclidean.py: fix the module docstring pointer to nonexistent
  jax_migration.md -> docs/user-guide/batching-jit.md
- decomposition: export lift_ideals and orthonormalize_rows (public-
  quality HoroPCA primitives already imported directly by tests)

Verified: test_helpers, test_gyro_normalization, test_horopca green;
ruff clean; imports verified from a fresh interpreter.
